Fix className for SVG elements
